+++ app/[artistAddress]/page.tsx
@@ -3,8 +3,8 @@ import ArtistPage from "@/components/ArtistPage";
 import ProfileProvider from "@/providers/ProfileProvider";
 import { APP_URL, VERCEL_OG } from "@/lib/og/consts";
 import truncateAddress from "@/lib/truncateAddress";
-import fetchArtistProfile from "@/lib/fetchArtistProfile";
 import { Address } from "viem";
+import getArtistProfile from "@/lib/getArtistProfile";
 
 type Props = {
   params: Promise<{ artistAddress: string }>;
@@ -13,7 +13,7 @@ type Props = {
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
   const { artistAddress } = await params;
 
-  const profile = await fetchArtistProfile(artistAddress as Address, true);
+  const profile = await getArtistProfile(artistAddress as Address);
   const title = profile.username || truncateAddress(artistAddress);
   const description = profile.bio || "Imagined by LATASHÁ";
